+++ app/src/main/java/pt/rvcoding/cvnotes/ui/profession/ProfessionState.kt
@@ -1,0 +1,5 @@
+package pt.rvcoding.cvnotes.ui.profession
+
+data class ProfessionState(
+    val profession: String
+)